Clean up Combobox more.
Moved query for the embedded <d-list>
object to createdCallback(), so Combobox can get a reference to the
real list (rather than place holder empty List) before
refreshRendering() runs.

One issue is that Combobox#_initValue() needs to be called every time the list is updated,
but it isn't.  Seems like a bug in the original code, and the test was previously just getting
lucky (due to a timing condition about when the list is updated vs. when Combobox first
looks at the list).  Rather than fixing it, I just changed the test.