Fix spurious test failure in Combobox unit test.
The timing in the previous version was too brittle.  It created the Comboboxes
in the beforeEach() preamble, and then called on("query-success", ...) in the main test
block, assuming that the "query-success" event had not yet fired.  But (perhaps due to
the Intern 3 upgrade) the event was firing before the main test block started.

Also, the main test block was strangley called "Store.add/remove/put" even though
it wasn't doing any of those things.

Also did some general cleanup.