Implement filterable multiselect.
Previously this wasn't possible because multiselect moved focus to the dropdown
to avoid some JAWS issue.  But that doesn't seem to be a problem anymore.

As before, the user navigates between the options using the arrow keys,
toggles options using the SPACE key, and closes the dropdown using the ENTER key.
This means that searching is limited to a single word; a user cannot search for a
string with a space in it.

When filtering is enabled, there is no way to see which items are selected,
or even how many are selected, without opening the dropdown.  A subclass can
add an indicator to the template to show how many items are selected.

Other changes include:

* Rename searchPlaceHolder to placeholder.  Unclear why it previously had a special name.
* Remove declarations of properties already inherited from LitFormWidget and
  LitFormValueWidget.
* Update old API documentation.

More combobox.

Revert to old displayedValue behavior for non-filterable.
For filterable, never set displayedValue.